+++ bin/civicplus.js
@@ -1,0 +1,9 @@
+#!/usr/bin/env node
+
+// @flow
+'use strict'
+
+const commands = require('../lib/commands')
+const { TENANTS } = require('../lib/config')
+
+commands(TENANTS.CIVICPLUS)