+++ examples/api/directory/boom/index.js
@@ -1,11 +1,11 @@
 'use strict'
 
 // https://www.npmjs.com/package/boom
-const Boom = require('boom')
+const Boom = require('@hapi/boom')
 
-module.exports = function (request) {
-  throw new Boom(request.url.query.message || null, {
+module.exports = function(request) {
+  throw new Boom.Boom(request.url.query.message || null, {
     statusCode: request.url.query.status || 500,
-    data: request.url
+    data: request.url,
   })
 }